+++ src/SharedComponents/StyledTextField/StyledTextField .tsx
@@ -1,18 +1,51 @@
-import React from 'react'
-import TextField from '@mui/material/TextField'
-import { styled } from '@mui/material/styles'
+import React, { useState } from 'react';
+import TextField from '@mui/material/TextField';
+import { styled } from '@mui/material/styles';
 
-const StyledTextField = styled(TextField)(({}) => ({
+const StyledLabel = styled('label')({
+  fontWeight: 'bold',
+  fontSize: '13px',
+  paddingBottom: '12px',
+  marginBottom: '110px',
+});
+
+const StyledTextField = styled(TextField)({
   '& .MuiInputBase-root': {
-    padding: '7px',
+    width: '100%', // Assurez-vous que le width est bien défini à 100%
     height: '40px',
-    width: '100%'
+    '& .MuiInputBase-input::placeholder': {
+      color: '#5d5a68',
+    },
+
+    '& input': {
+      padding: '7px',
+      height: '100%',
+    },
   },
-  '& input::placeholder': {}
-}))
+});
 
-const CustomTextField = ({ label, placeholder, ...rest }) => {
-  return <StyledTextField label={label} placeholder={placeholder} {...rest} />
-}
 
-export default CustomTextField
+const CustomTextField = ({ label, name, correctValue, placeholder, ...rest }) => {
+  const [isErrorVisible, setIsErrorVisible] = useState(false);
+
+  const handleInputChange = (event : any) => {
+    const inputValue = event.target.value;
+    setIsErrorVisible(inputValue !== correctValue);
+  };
+
+  return (
+    <>
+      <StyledLabel htmlFor='name'>{label} </StyledLabel>
+      <StyledTextField
+        name={name}
+        onChange={handleInputChange}
+        placeholder={placeholder}
+        error={isErrorVisible}
+        helperText={isErrorVisible ? 'Incorrect value' : ''}
+        {...rest}
+      />
+    </>
+  );
+};
+
+export default CustomTextField;
