+++ src/pages/Contract/Flow/index.tsx
@@ -2,10 +2,11 @@ import React from 'react'
 import { useForm, Controller } from 'react-hook-form'
 import CustomTextField from '../../../SharedComponents/StyledTextField/StyledTextField '
 
+   
 function Index() {
   const { control, handleSubmit } = useForm()
 
-  const onSubmit = data => {
+  const onSubmit = (data: any) => {
     console.log(data)
   }
 
@@ -17,13 +18,12 @@ function Index() {
         rules={{ required: true, minLength: 7 }}
         render={({ field, fieldState }) => (
           <CustomTextField
-            placeholder='fullname'
+            correctValue={undefined} placeholder='fullname'
             label='Full Name'
             control={control}
             {...field}
             error={fieldState.invalid}
-            helperMessage='aaa'
-          />
+            helperMessage='aaa'          />
         )}
       />
 
@@ -34,3 +34,4 @@ function Index() {
 }
 
 export default Index
+
